- refactor: Changed Profile struct to use non-pointer string types and updated Age to int16 in backend/internal/database/models.go. - refactor: Modified CreateProfileParams, PatchProfileParams, and UpdateProfileParams types in backend/internal/database/query.sql.go to reflect new data types. - refactor: Adjusted ProfileDTO and related structures in backend/internal/models/dto/profile.go for accurate JSON bindings. - refactor: Updated ProfileService implementation in backend/internal/services/profile.go for consistency with new data structures. - chore: Updated SQL queries in sqlc/query.sql and schema.sql to accommodate changes in data types for age. - chore: Adjusted configuration in sqlc.yaml for seamless integration with the new schema.

+++ backend/internal/models/dto/profile.go
@@ -3,18 +3,18 @@ package dto
 type ProfileDTO struct {
 	Name     string `json:"name"`
 	LastName string `json:"last_name"`
-	Age      int    `json:"age"`
+	Age      int16  `json:"age"`
 	Id       int32  `json:"id"`
 }
 
 type NewProfileDTO struct {
 	Name     string `json:"name" binding:"required"`
 	LastName string `json:"last_name" binding:"required"`
-	Age      int    `json:"age" binding:"required,min=18,max=100"`
+	Age      int16  `json:"age" binding:"required,min=18,max=100"`
 }
 
 type PatchProfileDTO struct {
 	Name     *string `json:"name"`
 	LastName *string `json:"last_name"`
-	Age      *int    `json:"age" binding:"omitempty,min=18,max=100"`
+	Age      *int16  `json:"age" binding:"omitempty,min=18,max=100"`
 }
